requirements.txt
fastapi
uvicorn[standard]
stripe
python-dotenv
sqlalchemy
alembic
python-jose[cryptography]
bcrypt
pydantic[email]
httpx
